Rename file from View.js to QuestionPage.js
